+++ wow_dbc/CHANGELOG.md
@@ -24,6 +24,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * BREAKING: `PowerType` renamed to `Power`.
 * BREAKING: `Faction` renamed to `LfgFaction`.
 * BREAKING: `Language` renamed to `ClientLanguage`.
+* BREAKING: `Category` renamed to `ServerCategory`.
 
 ### Removed